Remove integer cast
It is unecessary and constrains the size which could lead to weird errors.

+++ src/graph.cpp
@@ -113,7 +113,7 @@ std::pair< std::vector<int>, int> graph::breadthFirstSearch(int start, int end)
 			break;
 
 		// Obtain all of the current node's neighbors. If it has no neighbors, skip to the next node in the queue
-		if (currentNode >= (int) adjList.size())
+		if (currentNode >= adjList.size())
 			continue;
 		std::map<int, vertex> neighbors = adjList[currentNode];
 		for (unsigned int i = 0; i < neighbors.size(); ++i)
